+++ lib/iex/errors.rb
@@ -1,1 +1,2 @@
 require_relative 'errors/symbol_not_found_error'
+require_relative 'errors/client_error'